整理代码以及增加查的接口(到dictionary)

+++ business/zenofsword/car/common/src/main/java/course/linkflower/link/oneframework/model/Dictionary.java
@@ -14,5 +14,5 @@ public class Dictionary {
     private String value;
     private String key;
     private String type;
-    private Long ordering;
+    private Byte ordering;
 }
